+++ demo/src/main/java/com/example/demo/Main.java
@@ -2,6 +2,7 @@ package com.example.demo;
 
 import com.example.demo.service.AirtableService;
 import com.example.demo.service.ContentService;
+import com.example.envUtils.DotenvConfig;
 import org.json.JSONException;
 
 import java.io.FileNotFoundException;
@@ -12,7 +13,7 @@ public class Main {
     public static void main(String[] args) throws IOException, JSONException {
         AirtableService airtableService = new AirtableService();
 
-        airtableService.getResponseList("Testimonials" );
+        airtableService.getResponseList("Testimonials", DotenvConfig.get("BASE_ID"),DotenvConfig.get("ACCESS_TOKEN"));
 
       //  String str = airtableService.getResponse("Testimonial");
     }
