+++ demo/src/main/java/com/example/demo/service/IAirtableService.java
@@ -14,7 +14,7 @@ public interface IAirtableService {
 
     //Methods
    // HttpGet request(String tableName) throws JSONException, IOException, NoSuchAlgorithmException;
-    List<JSONObject> getResponseList(String tableName) throws JSONException, IOException;
+    List<JSONObject> getResponseList(String tableName,String baseId,String accessToken) throws JSONException, IOException;
     /*
     JSONObject getResponse(String tableName, HttpGet request) throws JSONException, NoSuchAlgorithmException, IOException;
